Move proptypes into main react-jw-player.jsx file
